refactor(#741): refactor caveController
