+++ api/controllers/v1/cave/set-massif.js
@@ -1,0 +1,45 @@
+const RightService = require('../../../services/RightService');
+
+const { checkIfExists } = sails.helpers;
+const { checkRight } = sails.helpers;
+
+module.exports = async (req, res) => {
+  // Check right
+  const hasRight = await checkRight
+    .with({
+      groups: req.token.groups,
+      rightEntity: RightService.RightEntities.CAVE,
+      rightAction: RightService.RightActions.EDIT_ANY,
+    })
+    .intercept('rightNotFound', () =>
+      res.serverError(
+        'A server error occured when checking your right to add a cave to a massif.'
+      )
+    );
+  if (!hasRight) {
+    return res.forbidden('You are not authorized to add a cave to a massif.');
+  }
+
+  // Check params
+  const caveId = req.param('caveId');
+  if (!(await checkIfExists('id', caveId, TCave))) {
+    return res
+      .status(404)
+      .send({ message: `Cave with id ${caveId} not found.` });
+  }
+
+  const massifId = req.param('massifId');
+  if (!(await checkIfExists('id', massifId, TMassif))) {
+    return res
+      .status(404)
+      .send({ message: `Massif with id ${massifId} not found.` });
+  }
+
+  // Update cave
+  await TCave.updateOne({
+    id: caveId,
+  }).set({
+    massif: massifId,
+  });
+  return res.sendStatus(204);
+};
